check semantics of lambda expression
add lambda_instantiation_map to semantics context, which preserves the
relation between lambda's type and the nodes which is actually emitted
by emitter.  This is needed because lambda type is treated as struct
type to capture variables temporarily.  So, on re-implementing lambda
object, I should consider about the instantiation map and how to treat
lambda object. (All lambda object may ought to be treated with the
instantiation map.)