change switch statement/expression grammar
because when ... grammar is not available if the statement is nested.
For example,

  func main
    when true
    when false
    else
    end
  end

If we want to parse this correctly, parser needs to know whole tokens.
When parser see the second 'when' token, parser interprets it as a
beginning of new switch statement. Finally, parser finds out that 'end'
for 'func' does not exist and notices the second 'when' token is for the
second clause of first switch statement. For this behavior, we need to
introduce back tracking to parser and it is not possible with LALR(1)
parsing. And even if introducing back tracking, the code would be
ambiguous when two switch statements are included in the block (which
switch statement should parser track back?).

Change:

From:
  when ...
    ...
  else
    ...
  end

To:
  switch
  case ...
    ...
  else
    ...
  end